# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Company Service",
    "version": "1.0.0",
    "description": "Documentation des routes"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Documentation Company Service"
    }
  ],
  "paths": {
    "/companies": {
      "get": {
        "tags": [
          "Company-Controller"
        ],
        "summary": "Get all companies data",
        "responses": {
          "200": {
            "description": "The company list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Notification"
                    },
                    "message": {
                      "type": "string",
                      "description": "The Company {name} was succesfull added"
                    },
                    "numberCompanies": {
                      "type": "number",
                      "description": "The number of companies"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "description": "The company's name."
                        },
                        "sector": {
                          "type": "string",
                          "description": "The company's sector."
                        },
                        "siren": {
                          "type": "number",
                          "description": "The company's siren."
                        },
                        "results": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "ca": {
                                "type": "number",
                                "description": "The company's CA."
                              },
                              "margin": {
                                "type": "number",
                                "description": "The company's margin."
                              },
                              "ebitda": {
                                "type": "number",
                                "description": "The company's ebitda."
                              },
                              "loss": {
                                "type": "number",
                                "description": "The company's loss."
                              },
                              "year": {
                                "type": "number",
                                "description": "The company's year."
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No company found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false,
                      "description": "Error"
                    },
                    "message": {
                      "type": "string",
                      "description": "No Company found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/companies/pages": {
      "get": {
        "tags": [
          "Company-Controller"
        ],
        "summary": "Get companies data by 10",
        "responses": {
          "200": {
            "description": "The company list by 10.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Notification"
                    },
                    "message": {
                      "type": "string",
                      "description": "The Company {name} was succesfull added"
                    },
                    "numberCompanies": {
                      "type": "number",
                      "description": "The number of companies"
                    },
                    "pageCount": {
                      "type": "number",
                      "description": "The number of pages (numberCompanies / 10)"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "description": "The company's name."
                        },
                        "sector": {
                          "type": "string",
                          "description": "The company's sector."
                        },
                        "siren": {
                          "type": "number",
                          "description": "The company's siren."
                        },
                        "results": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "ca": {
                                "type": "number",
                                "description": "The company's CA."
                              },
                              "margin": {
                                "type": "number",
                                "description": "The company's margin."
                              },
                              "ebitda": {
                                "type": "number",
                                "description": "The company's ebitda."
                              },
                              "loss": {
                                "type": "number",
                                "description": "The company's loss."
                              },
                              "year": {
                                "type": "number",
                                "description": "The company's year."
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No company found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false,
                      "description": "Error"
                    },
                    "message": {
                      "type": "string",
                      "description": "No Company found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}